- tentando corigir. - adicionando logs mais completos.

+++ src/back/Estimator.java
@@ -5,7 +5,7 @@ package back;
  */
 
 public interface Estimator {
-	int estimate(double collision_slots, double empty_slots, double successful_slots, double current_size, int[] frame);
+	double estimate(double collision_slots, double empty_slots, double successful_slots, double current_size, int[] frame);
 
 	String getName();
